# src/main/java/com/gshelgaas/bankcards/dto/BlockRequestDto.java
package com.gshelgaas.bankcards.dto;

import jakarta.validation.constraints.NotBlank;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class BlockRequestDto {
    @NotBlank(message = "Reason cannot be blank")
    private String reason;
}